% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R, R/ui_utils.R
\name{makeCard}
\alias{makeCard}
\title{Makes a simple card for a Shiny app}
\usage{
makeCard(title, content, size = 12, style = "")

makeCard(title, content, size = 12, style = "")
}
\arguments{
\item{title}{card title}

\item{content}{page contents}

\item{size}{text size}

\item{style}{style}
}
\value{
the card

the card
}
\description{
Makes a simple card for a Shiny app

Makes a simple card for a Shiny app
}
