% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R, R/ui_utils.R
\name{makePage}
\alias{makePage}
\title{Makes a simple page for a Shiny app}
\usage{
makePage(title, subtitle, contents)

makePage(title, subtitle, contents)
}
\arguments{
\item{title}{page title}

\item{subtitle}{page subtitle}

\item{contents}{page contents}
}
\value{
the page

the page
}
\description{
Makes a simple page for a Shiny app

Makes a simple page for a Shiny app
}
