% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_navigation.R
\name{navigation}
\alias{navigation}
\title{Title}
\usage{
navigation()
}
\value{
a navigation bar
}
\description{
For diffent icons search microsoft Fluent UI: https://developer.microsoft.com/en-us/fluentui#/styles/web/icons
}
